--- src/contexts/ThemeContext.jsx
@@ -1,19 +1,0 @@
-import React, { useState } from "react";
-
-const ThemeContext = React.createContext(null);
-
-function ThemeContextProvider(props) {
-  const [theme, setTheme] = useState("light");
-
-  function toggleTheme() {
-    setTheme((prev) => (prev === "light" ? "dark" : "light"));
-  }
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      {props.children}
-    </ThemeContext.Provider>
-  );
-}
-
-export { ThemeContextProvider, ThemeContext };
