+++ src/contexts/ThemeContext.tsx
@@ -1,0 +1,30 @@
+import React, { useState } from "react";
+
+type ThemeContextProviderProps = {
+  children: React.ReactNode;
+};
+
+type Theme = "light" | "dark";
+
+type ThemeContextType = {
+  theme: Theme;
+  toggleTheme: () => void;
+};
+
+const ThemeContext = React.createContext<ThemeContextType | null>(null);
+
+function ThemeContextProvider({ children }: ThemeContextProviderProps) {
+  const [theme, setTheme] = useState<Theme>("light");
+
+  function toggleTheme() {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  }
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+}
+
+export { ThemeContextProvider, ThemeContext };
